+++ Microsoft.Azure.Cosmos/src/Query/Core/Pipeline/CrossPartition/IQueryDataSource.cs
@@ -1,0 +1,35 @@
+﻿// ------------------------------------------------------------
+// Copyright (c) Microsoft Corporation.  All rights reserved.
+// ------------------------------------------------------------
+
+namespace Microsoft.Azure.Cosmos.Query.Core.Pipeline.CrossPartition
+{
+    using System.Threading;
+    using System.Threading.Tasks;
+    using Microsoft.Azure.Cosmos.Query.Core;
+    using Microsoft.Azure.Cosmos.Query.Core.Pipeline;
+
+    internal interface IQueryDataSource : IMonadicQueryDataSource
+    {
+        Task<QueryPage> QueryAsync(
+            SqlQuerySpec sqlQuerySpec,
+            string continuationToken,
+            Cosmos.PartitionKey partitionKey,
+            int pageSize,
+            CancellationToken cancellationToken);
+
+        Task<QueryPage> QueryAsync(
+            SqlQuerySpec sqlQuerySpec,
+            string continuationToken,
+            int partitionKeyRangeId,
+            int pageSize,
+            CancellationToken cancellationToken);
+
+        Task<QueryPage> QueryAsync(
+            SqlQuerySpec sqlQuerySpec,
+            string continuationToken,
+            FeedRangeInternal feedRange,
+            int pageSize,
+            CancellationToken cancellationToken);
+    }
+}